// client/packages/image-annotator/src/app.types.tsx
export type Maybe<T> = T | null | undefined;
// ...
export interface ImageSize {
    width: number;
    height: number;
}